Adds redirect module for legacy page urls
